25. Processamento de dados em Java, Casting

+++ Aula25/src/Aula25.java
@@ -48,7 +48,7 @@ public class Aula25 {
 		a = 5;
 		b = 2;
 		
-		resultado = a / b;
+		resultado = (double)a / b;
 		
 		System.out.println(resultado);
